합본2
컴포넌트 분리해놓음
