+++ src/components/markSelectForm/MarkSelectForm.css
@@ -1,0 +1,82 @@
+.markPage{
+    font-family: "Pretendard";
+}
+
+.littleTitle{
+    font-size: 26px;
+    text-align: left;
+    font-weight: 500;
+    color: #FFFFFF;
+}
+
+.littleInfo{
+    font-size: 16px;
+    text-align: left;
+    font-weight: 300;
+    color: #000000;
+    margin-bottom: 50px;
+}
+
+.contentName_row {
+    background-color: #3E3E3F;
+    font-size: 20px;
+    font-weight: 400;
+    text-align: center;
+    color: white;
+}
+
+.nationButton{
+    font-size: 16px;
+    font-weight: 400;
+    margin-left: 30px;
+    border-radius: 50px;
+    border: 0.5px solid #2F2E41;
+    background-color: transparent;
+    color: black; 
+    width: 120px;
+    height: 48px;
+    padding: 8px 16px;
+}
+.nationButton_table{
+    font-size: 12px;
+    font-weight: 400;
+    margin-left: 5px;
+    border-radius: 40px;
+    border: 0.5px solid #2F2E41;
+    background-color: transparent;
+    color: black; 
+    width: 100px;
+    height: 35px;
+    padding: 4px 8px;
+}
+
+.nationBox_selected{
+    background-color: #CBA585;
+    color: white;
+    border-radius: 30px;
+    border: 0.2px solid #CBA585;
+    width: 120px;
+    height: 25px;
+    margin-top: 10px;
+    text-align: center;
+    padding: 8px;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.5);
+}
+
+.confireButton_table{
+    font-size: 16px;
+    font-weight: 400;
+    border-radius: 50px;
+    border: 0.5px solid #CBA585;
+    background-color: white;
+    color: #CBA585; 
+    width: 120px;
+    height: 48px;
+    padding: 8px 16px;
+}
+
+.markType{
+    background-color: #3E3E3F;
+    margin-top: 130px;
+    min-height: 700px;
+}
